+++ src/Ui/Bundle/EventBundle/Handler/Catalog/CatalogFilterViewsHandler.php
@@ -18,6 +18,7 @@ use Proximum\Vimeet\Application\Query\Catalog\PositionViewQuery;
 use Proximum\Vimeet\Application\Query\Catalog\SearchFacet\SearchFacetViewQuery;
 use Proximum\Vimeet\Application\Query\Catalog\TypeViewQuery;
 use Proximum\Vimeet\Application\View\Catalog\SearchFacetsView;
+use Proximum\Vimeet\Domain\Catalog\GetDisplayObjectiveFilter;
 use Proximum\Vimeet\Domain\Catalog\VisibleParticipationCategories;
 use Proximum\Vimeet\Domain\Catalog\VisibleParticipationTypes;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
@@ -36,22 +37,28 @@ class CatalogFilterViewsHandler
     /** @var EngineInterface */
     private $engine;
 
+    /** @var GetDisplayObjectiveFilter */
+    private $getDisplayObjectiveFilter;
+
     /**
      * @param QueryBusInterface              $queryBus
      * @param VisibleParticipationCategories $visibleParticipationCategories
      * @param VisibleParticipationTypes      $visibleParticipationTypes
      * @param EngineInterface                $engine
+     * @param GetDisplayObjectiveFilter      $getDisplayObjectiveFilter
      */
     public function __construct(
         QueryBusInterface $queryBus,
         VisibleParticipationCategories $visibleParticipationCategories,
         VisibleParticipationTypes $visibleParticipationTypes,
-        EngineInterface $engine
+        EngineInterface $engine,
+        GetDisplayObjectiveFilter $getDisplayObjectiveFilter
     ) {
         $this->queryBus = $queryBus;
         $this->visibleParticipationCategories = $visibleParticipationCategories;
         $this->visibleParticipationTypes = $visibleParticipationTypes;
         $this->engine = $engine;
+        $this->getDisplayObjectiveFilter = $getDisplayObjectiveFilter;
     }
 
     /**
@@ -93,7 +100,8 @@ class CatalogFilterViewsHandler
                     $this->engine->renderResponse(
                         'EventBundle:Catalog:no-visible-category.html.twig',
                         ['event' => $event, 'sheet' => $sheet]
-                    )
+                    ),
+                    ($this->getDisplayObjectiveFilter)($sheet, $locale)
                 );
             }
 
@@ -112,7 +120,8 @@ class CatalogFilterViewsHandler
                     $this->engine->renderResponse(
                         'EventBundle:Catalog:no-visible-type.html.twig',
                         ['event' => $event, 'sheet' => $sheet]
-                    )
+                    ),
+                    ($this->getDisplayObjectiveFilter)($sheet, $locale)
                 );
             }
 
@@ -141,7 +150,9 @@ class CatalogFilterViewsHandler
             $typeViews,
             $organizationCategoryViews,
             $positionViews,
-            $taggedNomenclatureTagViews
+            $taggedNomenclatureTagViews,
+            null,
+            ($this->getDisplayObjectiveFilter)($sheet, $locale)
         );
     }
 }