+++ tests/Ui/Bundle/EventBundle/Handler/Catalog/CatalogFilterViewsHandlerTest.php
@@ -22,6 +22,7 @@ use Proximum\Vimeet\Application\Query\Catalog\TypeViewQuery;
 use Proximum\Vimeet\Application\View\Catalog\PositionView;
 use Proximum\Vimeet\Application\View\Catalog\SearchFacetsView;
 use Proximum\Vimeet\Application\View\Catalog\SearchFacetView;
+use Proximum\Vimeet\Domain\Catalog\GetDisplayObjectiveFilter;
 use Proximum\Vimeet\Domain\Catalog\VisibleParticipationCategories;
 use Proximum\Vimeet\Domain\Catalog\VisibleParticipationTypes;
 use Proximum\Vimeet\Domain\Model\Event;
@@ -57,6 +58,9 @@ class CatalogFilterViewsHandlerTest extends TestCase
     /** @var string */
     public $locale;
 
+    /** @var ObjectProphecy */
+    public $canDisplayObjectiveFilter;
+
     public function setUp()
     {
         $this->queryBus = $this->prophesize(QueryBusInterface::class);
@@ -66,6 +70,7 @@ class CatalogFilterViewsHandlerTest extends TestCase
         $this->event = $this->prophesize(Event::class);
         $this->sheet = $this->prophesize(Sheet::class);
         $this->locale = 'fr';
+        $this->canDisplayObjectiveFilter = $this->prophesize(GetDisplayObjectiveFilter::class);
     }
 
     public function testHandleNoType()
@@ -75,6 +80,8 @@ class CatalogFilterViewsHandlerTest extends TestCase
             $this->sheet->reveal(),
             $this->locale
         );
+
+        $this->canDisplayObjectiveFilter->__invoke($this->sheet->reveal(), $this->locale)->shouldBeCalled()->willReturn([]);
 
         $searchFacetView = new SearchFacetView('type', 'label', 'placeholder', true);
         $searchFacetsView = new SearchFacetsView([$searchFacetView]);
@@ -98,7 +105,8 @@ class CatalogFilterViewsHandlerTest extends TestCase
             $this->queryBus->reveal(),
             $this->visibleParticipationCategories->reveal(),
             $this->visibleParticipationTypes->reveal(),
-            $this->engine->reveal()
+            $this->engine->reveal(),
+            $this->canDisplayObjectiveFilter->reveal()
         );
 
         $result = $handler->handle($view);
@@ -123,6 +131,8 @@ class CatalogFilterViewsHandlerTest extends TestCase
             $this->sheet->reveal(),
             $this->locale
         );
+
+        $this->canDisplayObjectiveFilter->__invoke($this->sheet->reveal(), $this->locale)->shouldBeCalled()->willReturn(['need']);
 
         $searchFacetView1 = new SearchFacetView('type', 'label', 'placeholder', true);
         $searchFacetView2 = new SearchFacetView('position', 'label', 'placeholder', true);
@@ -174,7 +184,8 @@ class CatalogFilterViewsHandlerTest extends TestCase
             $this->queryBus->reveal(),
             $this->visibleParticipationCategories->reveal(),
             $this->visibleParticipationTypes->reveal(),
-            $this->engine->reveal()
+            $this->engine->reveal(),
+            $this->canDisplayObjectiveFilter->reveal()
         );
 
         $result = $handler->handle($view);
@@ -188,7 +199,8 @@ class CatalogFilterViewsHandlerTest extends TestCase
             [
                 'sheet_organization_category' => [],
             ],
-            null
+            null,
+            ['need']
         );
 
         $this->assertEquals($expected, $result);